Wake up TransactService when MMS data is available
In 2G network, mobile data is suspend during voice
call. MMS data will be notified as unavailable.
After voice call is end, It may notify Default data
connection before MMS data connection. When it
notify Default connection, it is unavailable in
MMS networkInfo and TransactionService will not
launch transaction; When it nofity MMS connection,
it is not connected so it will not wake up
Transaction Service.

How to reproduce:
1) Make a voice call in 2G network;
2) Try to send a MMS during voice call;
3) MMS is sending and waiting for data connection;
4) Hang up the voice call;
5) MMS stays in Sending state.

Change-Id: Ia719e8736cc6ccfcd236bceea4cb94f7c4dd94e5
Signed-off-by: Bin Li <libin@marvell.com>