Option to strip all unicode characters from SMS - CM
According to the GSM standard 03.38, SMS can contain limited number of
characters with accents (http://goo.gl/nBS99). This is why there is the
option "Strip unicode" which substitutes all unsupported characters by
ASCII equivalent (e.g. 'á' => 'a').

This patch set implements an option which allows to replace all accented
characters by their ASCII equivalent which makes the final message
better readable.

For example if only the "Strip unicode" option is enabled, the word
"zatažené" is changed to "zatazené". With the new option (Strip all
unicode characters), the same word is changed to "zatazene" which makes
it more naturally readable. In general it is easier to read text without
any accents than with just some of them.

Change-Id: I673bdfda5ff07e2e783bd97baa1e38ca84acd58e

Conflicts:
	res/values/arrays.xml
	res/xml/preferences.xml